requirements.txt
amqp==5.3.1
APScheduler==3.10.4
asgiref==3.7.2
billiard==4.2.1
boto3==1.34.34
botocore==1.34.34
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==4.2.7
django-autoslug==1.9.9
django-celery-beat==2.7.0
django-cors-headers==4.3.1
django-jazzmin==2.6.0
django-js-asset==2.2.0
django-phonenumber-field==7.3.0
django-redis==5.4.0
django-storages==1.14.2
django-timezone-field==7.0
django-tinymce==3.6.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
jmespath==1.0.1
kombu==5.4.2
packaging==23.2
phonenumberslite==8.13.28
pillow==10.2.0
prompt_toolkit==3.0.48
psycopg2==2.9.10
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.2.0
s3transfer==0.10.0
setuptools==75.2.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.13
